iconGrid: Ensure minimum item size is big enough for label and icon
The minimum item size was using a fixed size of MIN_ICON_SIZE, which is
16px. This did not consider the label or padding added by app grid
items. So such a minimum size would often be smaller the space required
by label and padding alone, resulting in negative icon sizes after icon
scaling when there is not much room for icons.

Another contributing factor to this problem was that the scale factor
was not considered, since (fixed)ItemSize is supposed to be scaled.

https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/1420