windowManager: Animate the appearance of new windows after overview closes
So that the animation doesn't get skipped if the overview is visible but
in the process of closing.

In this case the overview being visible also means that it is closing,
because if it's visible and not already closing then `_shouldAnimateActor`
would have returned false and exited `_mapWindow` earlier.

Fixes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/1814,
       https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/3242

(cherry picked from commit 88b2c9a046170d150e41f929810b8fa9f5198ca6)

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/1982>