userWidget: Disconnect signal connections on destroy
At the moment an Avatar (and StIcon) are getting leaked
every time the screen is locked.

This is because the Avatar class sets up a global signal
connection that prevents the object from getting garbage
collected when its actor is destroyed.

This commit adds a destroy handler to disconnect the
signal connection.