screenshot: Grab screenshot during paint on X11
We might not be able to directly paint the stage to an offscreen, if
there currently is a fullscreen unredirected window. To make it possible
to take a screenshot in this situation, disable unredirecting, queue a
frame, and take a screenshot after having painted that frame, before we
go back being unredirected.

Don't do this on Wayland because it's a waste.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/1453


(cherry picked from commit c09be8b0a9395f5ce0c4210356f6532bd3bfa0c4)

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/1536>