viewSelector: Move showApps() to Overview
Now that Overview is able to ease into any state, be it window
picker or app grid, we can move this ViewSelector method to
Overview itself, which is its rightful place to live.

Move ViewSelector.showApps() to Overview.showApps(), and adapt
ShellDBus to call it from overview. Also make sure the show apps
button is correctly toggled.